Updated for experiment 72, and made more batches per network.
